frontend1/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
openai==0.28.0
webrtcvad==2.0.10
pyaudio @ https://people.csail.mit.edu/hubert/pyaudio/packages/pyaudio-0.2.11-cp311-cp311-linux_x86_64.whl
python-dotenv==1.0.1
requests==2.32.3
sounddevice==0.5.1
soundfile==0.13.1
SpeechRecognition==3.14.2
numpy==2.1.3
pydantic==2.10.6
pydantic_core==2.27.2
gunicorn
